Move wheel building into its own module
